+++ QATCH/core/ring_buffer.py
@@ -2,9 +2,33 @@
 ring_buffer.py
 
 This module provides a high-performance ``RingBuffer`` class backed by NumPy.
-This buffer maintains a fixed-size backing array and an internal write head.
-This guarantees ``O(1)`` time complexity for data insertion, preventing process stalls
-or synchronization issues during tight data-ingestion loops.
+It maintains a contiguous linear arena and a moving write head, guaranteeing
+amortized ``O(1)`` insertion AND ``O(1)`` reads. This prevents process stalls or
+synchronization issues during tight data-ingestion loops, and -- critically --
+during the much hotter read path that drains the buffer many times per UI tick.
+
+Design history (why this is arena-backed rather than wrapped):
+    - A ``np.roll`` design is ``O(N)`` on every append (it copies the whole
+      array each insertion), which stalls tight ingestion loops.
+    - A wrapped head/tail design fixes append to ``O(1)`` but forces reads to
+      ``np.concatenate`` the two halves once the buffer wraps -- a fresh
+      ``O(N)`` allocation on every ``get_partial``/``get_all`` call. Because the
+      render path reads the buffer dozens of times per tick (per channel, per
+      curve, per detector), that merely relocates the ``O(N)`` cost from the
+      producer to the hotter consumer side.
+    - This arena design keeps the live samples ALWAYS contiguous in a backing
+      array sized ``2 * size_max``. Appends advance a tail; the live window is
+      slid back to the front only when the tail reaches the end of the arena
+      (at most once per ``size_max`` appends -> amortized ``O(1)``). Reads then
+      return a plain contiguous slice -- a view, with no allocation or copy.
+
+Consumer contract:
+    ``get_partial`` and ``get_all`` (when full) return a VIEW into the internal
+    backing store for speed, not a snapshot. Consume the result (slice it, copy
+    it, feed it to ``setData`` / ``np.average`` / ``np.convolve``) before the
+    next ``append`` on the same buffer. Do not retain a returned array across
+    appends and expect it to stay frozen. All current callers read-then-discard
+    within a single tick, which is safe.
 
 References:
     - http://code.activestate.com/recipes/68429-ring-buffer/
@@ -30,15 +54,17 @@ warnings.filterwarnings("ignore", category=RuntimeWarning)
 
 
 class RingBuffer:
-    """An ``O(1)`` circular-buffer implementation backed by a numpy array.
+    """An amortized-``O(1)`` circular buffer backed by a contiguous numpy arena.
 
-    This replaces previous `np.roll`-based designs to avoid ``O(N)`` copy operations
-    per append. It maintains a fixed-size backing array and a write head, never
-    reallocating after construction.
+    The backing array has room for ``2 * size_max`` elements so that the live
+    sample window can grow rightward before being slid back to the front. The
+    live region (``[_start:_end]``) is always contiguous, so reads return a
+    zero-copy view. See the module docstring for the consumer contract regarding
+    returned views.
 
     Attributes:
-        size_max (int): The maximum capacity of the buffer.
-        size (int): The current number of valid samples written to the buffer.
+        size_max (int): The maximum capacity (window length) of the buffer.
+        size (int): The current number of valid samples held in the window.
     """
 
     def __init__(self, size_max: int, default_value: Any = 0, dtype: npt.DTypeLike = float) -> None:
@@ -46,7 +72,7 @@ class RingBuffer:
 
         Args:
             size_max (int): The maximum number of elements the buffer can hold.
-            default_value (Any, optional): The value used to pre-fill the buffer. Defaults to 0.
+            default_value (Any, optional): The value used to pre-fill the arena. Defaults to 0.
             dtype (npt.DTypeLike, optional): The numpy data type of the array. Defaults to float.
 
         Raises:
@@ -62,82 +88,103 @@ class RingBuffer:
         self._default_value: Any = default_value
         self._dtype: npt.DTypeLike = dtype
 
-        # NOTE: _head points at the slot the NEXT append will write into.
-        # NOTE: The most-recently-written sample is therefore at _head - 1.
-        self._head: int = 0
-        self._full: bool = False
+        # The arena holds 2x the window. _start indexes the oldest live sample;
+        # _end is one past the newest. The live region is always contiguous and
+        # satisfies the invariant: _end - _start == size.
+        self._cap: int = 2 * size_max
+        self._start: int = 0
+        self._end: int = 0
         self.size: int = 0
-        self._data: np.ndarray = np.full(size_max, default_value, dtype=dtype)
+        self._data: np.ndarray = np.full(self._cap, default_value, dtype=dtype)
 
     def append(self, value: Any) -> None:
-        """Appends new data to the ring buffer in ``O(1)`` time.
+        """Appends new data to the ring buffer in amortized ``O(1)`` time.
 
-        If the buffer is full, the oldest sample is overwritten.
+        If the window is full, the oldest sample is dropped by advancing the
+        window start. No per-append reallocation or full-array copy occurs; the
+        live window is slid back to the front of the arena at most once every
+        ``size_max`` appends.
 
         Args:
             value (Any): The value to append to the buffer.
         """
-        self._data[self._head] = value
-        self._head = (self._head + 1) % self.size_max
+        if self._end >= self._cap:
+            # Slide the live window back to the front. The source and
+            # destination ranges do not overlap (cap == 2 * size_max), so a
+            # plain block copy is safe. Amortized O(1) over size_max appends.
+            np.copyto(self._data[: self.size], self._data[self._start : self._end])
+            self._start = 0
+            self._end = self.size
 
-        if not self._full:
+        self._data[self._end] = value
+        self._end += 1
+
+        if self.size < self.size_max:
             self.size += 1
-            if self.size == self.size_max:
-                self._full = True
+        else:
+            # Window already full: drop the oldest sample.
+            self._start += 1
 
     def get_all(self) -> np.ndarray:
-        """Returns all elements from the buffer, preserving capacity size.
+        """Returns the buffer contents from oldest to newest.
 
-        When the buffer is not yet full, it returns the written chronological
-        region followed by default-filled values to preserve the old contract.
+        When the buffer is full, this returns a zero-copy VIEW of the live
+        window (see the module-level consumer contract). When it is not yet
+        full, it returns a freshly allocated ``size_max``-length array with the
+        written samples followed by ``default_value`` padding, preserving the
+        legacy capacity-sized contract.
 
         Returns:
-            np.ndarray: A copy of the array from oldest to newest.
+            np.ndarray: Elements ordered oldest to newest.
         """
-        if not self._full:
-            return self._data.copy()
+        if self.size >= self.size_max:
+            return self._data[self._start : self._end]
 
-        # Oldest sample sits at _head, newest at _head - 1.
-        return np.concatenate((self._data[self._head :], self._data[: self._head]))
+        out = np.full(self.size_max, self._default_value, dtype=self._dtype)
+        out[: self.size] = self._data[self._start : self._end]
+        return out
 
     def get_partial(self) -> np.ndarray:
         """Returns only the valid, written samples in chronological order.
 
-        If no valid samples exist ``(size == 0)``, returns the default-filled array.
+        This is the hot-path accessor. When at least one sample is present it
+        returns a zero-copy VIEW of the live window (oldest to newest); see the
+        module-level consumer contract. If no valid samples exist
+        ``(size == 0)`` it returns the default-filled capacity array.
 
         Returns:
-            np.ndarray: A slice of the buffer containing only written data from oldest to newest.
+            np.ndarray: The valid samples from oldest to newest.
         """
-        if self.size == 0 or self._full:
+        if self.size == 0:
             return self.get_all()
 
-        return self._data[: self.size].copy()
+        return self._data[self._start : self._end]
 
     def get_newest(self) -> Any:
         """Retrieves the most recently added element in the buffer.
 
         Raises:
-            IndexError: If the buffer is empty.
+            ValueError: If the buffer is empty.
 
         Returns:
             Any: The newest value.
         """
         if self.size == 0:
             raise ValueError("RingBuffer is empty")
-        return self.get_partial()[-1]
+        return self._data[self._end - 1]
 
     def get_oldest(self) -> Any:
         """Retrieves the oldest valid element remaining in the buffer.
 
         Raises:
-            IndexError: If the buffer is empty.
+            ValueError: If the buffer is empty.
 
         Returns:
             Any: The oldest value.
         """
         if self.size == 0:
             raise ValueError("RingBuffer is empty")
-        return self.get_partial()[0]
+        return self._data[self._start]
 
     def __getitem__(self, key: Union[int, slice]) -> Any:
         """Indexes into the chronological, oldest-to-newest view of the valid data.
