--- QATCH/nightly/.gitignore
@@ -1,2 +1,0 @@
-create_secret.py
-latest_build.json
