+++ QATCH/processors/nightly/.gitignore
@@ -1,0 +1,2 @@
+create_secret.py
+latest_build.json
